create routing not on conflict
